pyproject.toml
[project]
name = "seller-toolset"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datetime>=6.0",
    "flask>=3.1.2",
    "gunicorn>=25.0.1",
    "numpy>=2.4.2",
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
    "pillow>=12.1.1",
    "pypdf>=6.6.2",
    "python-barcode[images]>=0.16.1",
    "rembg>=2.0.72",
    "reportlab>=4.4.10",
]
